chg: refactor code for standard
