+++ ap/mock/products.js
@@ -1,0 +1,39 @@
+const products = [
+  {
+    id: 1,
+    name: "ball",
+    descr: "small sphere used in sports games",
+    price: 15000,
+    creationDate: "10/16/2024",
+  },
+  {
+    id: 2,
+    name: "frisbee",
+    descr: "flying disc used for recreational activities",
+    price: 15000,
+    creationDate: "10/16/2024",
+  },
+  {
+    id: 3,
+    name: "bicycle",
+    descr: "two-wheeled vehicle for transportation",
+    price: 350000,
+    creationDate: "10/16/2024",
+  },
+  {
+    id: 4,
+    name: "racket",
+    descr: "tool for playing tennis",
+    price: 80000,
+    creationDate: "10/16/2024",
+  },
+  {
+    id: 5,
+    name: "roller skates",
+    descr: "footwear with wheels for gliding",
+    price: 120000,
+    creationDate: "10/16/2024",
+  },
+];
+
+module.exports = products;
