+++ ap/models/product.js
@@ -1,11 +1,11 @@
 class Product {
-    constructor(id, name, descr, price, creationDate) {
-        this.id = id;
-        this.name = name;
-        this.descr = descr;
-        this.price = price;
-        this.creationDate = creationDate;
-    }
+  constructor(id, name, descr, price, creationDate) {
+    this.id = id;
+    this.name = name;
+    this.descr = descr;
+    this.price = price;
+    this.creationDate = creationDate;
+  }
 }
 
 module.exports = Product;
